+++ spec/disabling_posting_to_slack_spec.rb
@@ -4,9 +4,9 @@ require 'spec_helper'
 
 describe Slackistrano do
   context 'when :slackistrano is disabled' do
-    before(:all) { set :slackistrano, false }
+    before(:all) { set(:slackistrano, false) }
 
-    %w[updating reverting updated reverted failed].each do |stage|
+    %w[starting updating reverting updated reverted failed].each do |stage|
       it "doesn't post on slack:deploy:#{stage}" do
         expect_any_instance_of(Slackistrano::Capistrano).not_to(receive(:post))