+++ spec/dry_run_spec.rb
@@ -11,7 +11,7 @@ end
 describe Slackistrano do
   before(:all) { set :slackistrano, klass: DryRunMessaging }
 
-  %w[updating reverting updated reverted failed].each do |stage|
+  %w[starting updating reverting updated reverted failed].each do |stage|
     it "does not post to slack on slack:deploy:#{stage}" do
       allow_any_instance_of(Slackistrano::Capistrano).to(
         receive(:dry_run?).and_return(true)